pyproject.toml
[project]
name = "lambda-machine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fábio Macêdo Mendes", email = "fabiomacedomendes@gmail.com" },
]
requires-python = ">=3.10"
dependencies = ["lark-parser>=0.12.0"]

[project.scripts]
lambda-machine = "lambda_machine:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["hypothesis>=6.131.20", "pytest>=8.3.5", "ruff>=0.11.10"]
